# api_glewe/api_glewe/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " OpenApi Documentation",
        "description": " Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/offer/enrolment": {
            "post": {
                "tags": [
                    "enrolment-course"
                ],
                "operationId": "4a2d0c67bceee37167c37e482900e3b3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "example": "99487892-f338-4740-88e7-b3377eafe173"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "I am registering for a course"
                    }
                }
            }
        },
        "/api/offer/course": {
            "post": {
                "tags": [
                    "user-course"
                ],
                "operationId": "f4168f32d17a5b2de46c86c2c4a9758c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "99487892-f338-4740-88e7-b3377eafe173"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "my course"
                    }
                }
            }
        },
        "/api/offer/rating-course-by-user": {
            "post": {
                "tags": [
                    "rating-course"
                ],
                "operationId": "667b47f98f153965dac9b3d15fa2bf35",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "example": "99487892-f338-4740-88e7-b3377eafe173"
                                    },
                                    "note": {
                                        "type": "integer",
                                        "example": 12
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "I note the trainer"
                    }
                }
            }
        },
...
    }
}